refactor: 컴포넌트명 명시화 및 export module 수정
