+++ src/components/Calculator/index.tsx
@@ -1,29 +1,6 @@
-import { useCallback, useState } from 'react';
-
-import TotalPad from '@/components/Calculator/TotalPad';
-import { Operator } from '@/constants/operation';
-import { calculatorMachine } from '@/model/calculator';
-
-import ButtonSection from './ButtonSection';
-import styles from './index.module.css';
-
-export type CalculatorArgs = Operator | number | string | undefined;
-
-const Calculator = () => {
-  const [state, setState] = useState('0');
-  const handleClick = useCallback(
-    (args: CalculatorArgs) => () => {
-      setState((prev) => calculatorMachine(prev, args));
-    },
-    []
-  );
-
-  return (
-    <div className={styles.calculator}>
-      <TotalPad computedInput={String(state)} />
-      <ButtonSection onClick={handleClick} />
-    </div>
-  );
-};
-
-export default Calculator;
+export { default as AllClearButton } from '@/components/Calculator/AllClearButton';
+export { default as ButtonSection } from '@/components/Calculator/ButtonSection';
+export { default as Calculator } from '@/components/Calculator/Calculator';
+export { default as DigitButton } from '@/components/Calculator/DigitButton';
+export { default as OperationButton } from '@/components/Calculator/OperationButton';
+export { default as TotalPad } from '@/components/Calculator/TotalPad';
